Map implemented.  Currently the Settings button leads directly to the map with a hardcoded location.
GPS implemented to show current location as well.  Will not show until
you go to DDMS (Window > Change Perspective > Other > DDMS) and manually
input a fake GPS location (click the emulator at the top left section
then go down to the emulator window and click send under the GPS tab). 

May have to implement GPS in order for it to work.  Go to the AVD
manager or whatever and edit the API you are using.  Then on the bottom
section, just choose to add GPS Support. 

Also, I had to go to the Top Ten project properties and got to Android
and switch over to Google 2.1 to make this run.